+++ lib/multicloud/core.rb
@@ -7,6 +7,8 @@ require "multicloud/core/entities/file"
 require "multicloud/core/entities/directory"
 require "multicloud/core/entities/request"
 require "multicloud/core/entities/operation_result"
+require "multicloud/core/entities/async_result_wrapper"
+require "multicloud/core/entities/async_status"
 require "multicloud/core/operation"
 # Errors
 require "multicloud/core/errors"