+++ lib/multicloud/core/entities/operation_result.rb
@@ -14,10 +14,18 @@ module Multicloud
         attr_reader :operation
 
         def initialize(operation:, result:, success: true, error: nil)
+          @async = result.is_a?(AsyncResultWrapper)
           @operation = operation
           @success = success
-          @result = result
+          @result = @async ? result.url : result
           @error = error
+        end
+
+        # Returns true if result is an async.
+        # @since 0.4.0
+        # @return [Boolean]
+        def async?
+          @async
         end
       end
     end